fix(dsp): stop shifting negative offsets left on the Cortex-M4 packing path
Twelve offset-packing sites inside ARM_MATH_DSP blocks shifted a
possibly-negative quantization offset left:

    packed = (offset << 16U) | (offset & 0x0FFFFL);

The public headers document these offsets as ranging -127 to 128, so a
negative operand is reachable by contract and the shift is undefined.
Each site now builds the pair in unsigned arithmetic, per the idiom the
issue prescribes:

    packed = (int32_t)(((uint32_t)offset << 16) | ((uint32_t)offset & 0xFFFFu));

The issue names eight sites in the s8 add and sub kernels; the identical
pattern also lives at four sibling sites in the s8 mul kernels
(arm_elementwise_mul_s8.c, arm_mul_scalar_s8.c), fixed in the same sweep.

The seventy `value * (1 << left_shift)` sites #355 introduced across
fourteen files computed the scale with a signed shift that is itself
undefined at left_shift == 31, which no kernel validates. Per the
issue's prescription they become `value * (int32_t)((uint32_t)1 << left_shift)`,
defined for the full width. (At 31 the signed multiply can still
overflow for |value| > 1; bounding left_shift remains open on #357.)

Verification, since none of the packing sites compile on the x86 host:

- An x86 harness reproducing the exact old and new expressions shows
  them bit-identical for every offset in the full int16 range (65536
  values, a superset of the documented -127..128) against an
  independent bit-model reference, and for every (value, left_shift)
  pair with left_shift 0..30 whose product fits int32 (1179644 pairs).
- Under gcc -fsanitize=undefined -fno-sanitize-recover=all the old
  expressions trap ("left shift of negative value -1"; "left shift of
  1 by 31 places cannot be represented in type 'int'") and the new
  ones run clean, including left_shift == 31.
- Mechanically reverting every edit by regex reproduces origin/main
  byte-identically for all sixteen files (checked before clang-format
  rewrapped two lines in the squared_difference scalar kernels).
- All sixteen files compile clean on the host (-Wall -Wextra -Werror,
  ARM_MATH_DSP off) and parse clean with ARM_MATH_DSP forced on.

On-target execution comes from the CI m4 legs: the coverage leg (-O0)
and the shipped-flags leg (-Ofast) both take the ARM_MATH_DSP path.

The host-sanitizer shift-base mask is left untouched: PACK_Q15x2_32x1
and the two header helpers are inlined into TUs including several of
the files #355 cleaned, so a per-file narrowing cannot be enumerated
without an unmasked full-suite run; that stays open on #357.

Refs #357
Refs #346

+++ Source/BasicMathFunctions/arm_elementwise_sub_s16.c
@@ -115,9 +115,9 @@ arm_cmsis_nn_status arm_elementwise_sub_s16(const int16_t *input_1_vect,
         two_halfword_1 = arm_nn_read_q15x2_ia(&input_1_vect);
         two_halfword_2 = arm_nn_read_q15x2_ia(&input_2_vect);
 
-        input_1 = (int16_t)(two_halfword_1 & 0xFFFF) * (1 << left_shift);
+        input_1 = (int16_t)(two_halfword_1 & 0xFFFF) * (int32_t)((uint32_t)1 << left_shift);
         input_1 = arm_nn_requantize(input_1, input_1_mult, input_1_shift);
-        input_2 = (int16_t)(two_halfword_2 & 0xFFFF) * (1 << left_shift);
+        input_2 = (int16_t)(two_halfword_2 & 0xFFFF) * (int32_t)((uint32_t)1 << left_shift);
         input_2 = arm_nn_requantize(input_2, input_2_mult, input_2_shift);
         diff = input_1 - input_2;
         diff = arm_nn_requantize(diff, out_mult, out_shift);
@@ -125,9 +125,9 @@ arm_cmsis_nn_status arm_elementwise_sub_s16(const int16_t *input_1_vect,
         diff = MIN(diff, out_activation_max);
         diff_1 = (int16_t)diff;
 
-        input_1 = (int16_t)(two_halfword_1 >> 16) * (1 << left_shift);
+        input_1 = (int16_t)(two_halfword_1 >> 16) * (int32_t)((uint32_t)1 << left_shift);
         input_1 = arm_nn_requantize(input_1, input_1_mult, input_1_shift);
-        input_2 = (int16_t)(two_halfword_2 >> 16) * (1 << left_shift);
+        input_2 = (int16_t)(two_halfword_2 >> 16) * (int32_t)((uint32_t)1 << left_shift);
         input_2 = arm_nn_requantize(input_2, input_2_mult, input_2_shift);
         diff = input_1 - input_2;
         diff = arm_nn_requantize(diff, out_mult, out_shift);
@@ -144,8 +144,8 @@ arm_cmsis_nn_status arm_elementwise_sub_s16(const int16_t *input_1_vect,
     while (loop_count > 0)
     {
         /* C = A - B */
-        input_1 = *input_1_vect++ * (1 << left_shift);
-        input_2 = *input_2_vect++ * (1 << left_shift);
+        input_1 = *input_1_vect++ * (int32_t)((uint32_t)1 << left_shift);
+        input_2 = *input_2_vect++ * (int32_t)((uint32_t)1 << left_shift);
 
         input_1 = arm_nn_requantize(input_1, input_1_mult, input_1_shift);
         input_2 = arm_nn_requantize(input_2, input_2_mult, input_2_shift);
